fix: align SSR and wildcard shared-key matching
Reuse findSharedKey for SSR VM lookup so import:false parents skip
COMMON_SHARED_SUBPATHS and trailing-slash wildcards pick the longest
prefix. Match node_modules candidates on a path-segment or module
extension boundary, not a naive `.` prefix.
